+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+### Added
+- `gamedata::VOICE_OF_THE_FALLEN_ID` and `gamedata::CLAW_OF_THE_FALLEN_ID`.
+
 ### Fixed
 - Handling of log files with "Claw of the Fallen" as the encounter id.
 - Both bosses are now returned for the "Voice & Claw of the Fallen" strike